move server into separate package and use charsequence for some things
using charsequence allows me to cause minimal overhead from string allocations when the fields aren't used for anything
